+++ Anathema/Source/Tools/Results/IResultsMVP.cs
@@ -111,7 +111,7 @@ namespace Anathema
 
         public ListViewItem GetItemAt(Int32 Index)
         {
-            ListViewItem Item = ListViewCache.Get(Index);
+            ListViewItem Item = ListViewCache.Get((UInt64)Index);
 
             // Try to update and return the item if it is a valid item
             if (Item != null && ListViewCache.TryUpdateSubItem(Index, ValueIndex, Model.GetValueAtIndex(Index)))
